Create starter upper
I had a couple hours on our drive up to the beach to start working on this. So far this is just sort of an outline, but it works with the first set of commands up untill they get into the cave. Hopefully we can continue/edit/work with this. The only thing I had trouble with is getting the player to enter the cave without it spitting out "(first opening the cave entrance")